+++ backend/api/admin.py
@@ -5,7 +5,8 @@ from django.contrib.auth.forms import ReadOnlyPasswordHashField
 from django.contrib.auth.models import Group
 from django.core.exceptions import ValidationError
 
-from .models import *
+from api.models import Lecture, Problem
+from api.models.auth import User
 
 
 class UserCreationForm(forms.ModelForm):
@@ -69,4 +70,5 @@ admin.site.register(User, UserAdmin)
 admin.site.unregister(Group)
 
 admin.site.register(Lecture)
+
 admin.site.register(Problem)
