+++ backend/api/models/__init__.py
@@ -1,0 +1,6 @@
+from .auth import User
+from .enrollment import Enrollment
+from .lecture import Lecture
+from .problem import Problem
+from .storage import Storage
+from .submission import Submission, SubmissionState
