+++ backend/api/serializers/auth_serializers.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+
+from api.models import User
+
+
+class UserSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = User
+        fields = ("id", "student_id", "password")
