+++ backend/api/serializers/execute_serializers.py
@@ -1,0 +1,6 @@
+from rest_framework import serializers
+
+
+class ExecuteSerializer(serializers.Serializer):
+    input = serializers.CharField(required=False)
+    code = serializers.CharField()
