+++ backend/api/serializers/storage_serializers.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+
+from api.models import Storage
+
+
+class StorageSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Storage
+        fields = "__all__"
