+++ backend/api/serializers/submission_serializers.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+
+from api.models import Submission
+
+
+class SubmissionSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Submission
+        fields = "__all__"
