--- backend/api/views.py
@@ -1,191 +1,0 @@
-import uuid
-from datetime import datetime
-from os import remove
-
-from django.contrib.auth import authenticate, login as auth_login, logout as auth_logout
-from rest_framework import status
-from rest_framework.decorators import (
-    api_view,
-    authentication_classes,
-    permission_classes,
-)
-from rest_framework.request import Request
-from rest_framework.response import Response
-
-import api.executor as executor
-from server.settings.base import BASE_DIR
-from .serializers import *
-
-
-@api_view(["POST"])
-@authentication_classes([])
-@permission_classes([])
-def login(request: Request):
-    student_id = request.data.get("student_id")
-    password = request.data.get("password")
-
-    # 요청 validation
-    if student_id is None or password is None:
-        return Response({"error": "student_id or password is not given"}, status=400)
-
-    # 유저 정보 검증
-    user = authenticate(student_id=student_id, password=password)
-    if user is None:
-        return Response({"error": "student_id or password is wrong"}, status=400)
-
-    # 로그인 처리
-    auth_login(request, user)
-    return Response({"id": user.id, "student_id": user.student_id}, status=200)
-
-
-@api_view(["GET"])
-def logout(request: Request):
-    # 로그아웃 처리
-    auth_logout(request)
-    return Response({"success": "logout success"}, status=200)
-
-
-@api_view(["GET"])
-def get_lectures(request):
-    lectures = Lecture.objects.filter(deadline__gte=datetime.now())
-    serializer = LectureSerializer(lectures, many=True)
-
-    return Response(serializer.data)
-
-
-@api_view(["GET"])
-def get_lecture_by_id(request, id):
-    login_user_id = request.user.id  # user id
-
-    lecture = Lecture.objects.filter(id=id)
-    problems = Problem.objects.filter(lecture=id)
-    enroll = Enrollment.objects.filter(user_id=login_user_id)
-
-    serializer = OneLectureSerializer(lecture, many=True)
-    serializer2 = ProblemSerializer(problems, many=True)
-    serializer3 = EnrolledLectureSerializer(enroll, many=True)
-
-    enrolled_lectures = []
-    for i in serializer3.data:
-        enrolled_lectures.append(i["lecture_id"])
-    serializer.data[0]["problems"] = serializer2.data
-    serializer.data["Problem"] = serializer2.data
-
-    # enrollment check
-    if id in enrolled_lectures:
-        return Response(serializer.data)
-    else:
-        return Response("Invalid Access")
-
-
-@api_view(["POST"])
-def enroll_lecture(request, id):
-    enroll = Enrollment.objects.all()
-    lectures = Lecture.objects.filter(deadline__gte=datetime.now())
-
-    serializer = EnrolledLectureSerializer(enroll, many=True)
-    serializer2 = LectureSerializer(lectures, many=True)
-
-    lecture_list = []
-    for i in serializer2.data:
-        lecture_list.append(i["lecture_id"])
-
-    # check valid lecture
-    if id in lecture_list:
-
-        enrolled_lectures = []
-        for i in serializer.data:
-            enrolled_lectures.append(i["lecture_id"])
-
-        # non-exist
-        if id not in enrolled_lectures:
-            serializer3 = EnrolledLectureSerializer(data=request.data)
-            if serializer3.is_valid():
-                serializer3.save()
-                return Response(status=200)
-            return Response(serializer3.errors, status=status.HTTP_400_BAD_REQUEST)
-
-    else:
-        return Response("Invalid Lecture")
-
-
-@api_view(["GET"])
-def get_problem_by_id(request: Request, id):
-    # 문제, 저장소, 제출 가져오기
-    problem = Problem.objects.filter(id=id).prefetch_related("storage_set").prefetch_related("submission_set").get()
-    storages = problem.storage_set.all()
-    submissions = problem.submission_set.all()
-
-    problem_dict = ProblemSerializer(problem).data
-    problem_dict["storages"] = StorageSerializer(storages, many=True).data
-    problem_dict["submissions"] = StorageSerializer(submissions, many=True).data
-
-    return Response(problem_dict, status=200)
-
-
-@api_view(["POST"])
-def execute(request: Request):
-    code = request.data.get("code")
-    inp = request.data.get("input")
-
-    file_path = BASE_DIR.joinpath("tmp", f"{uuid.uuid4()}.py")
-
-    # create directory if not exists
-    if not file_path.parent.exists():
-        file_path.parent.mkdir()
-
-    # write code to file
-    with open(file_path, "w") as f:
-        f.write(code)
-
-    outp, err = executor.run(file_path, code, inp)
-
-    # remove file
-    remove(file_path)
-
-    return Response({"result": outp, "error": err})
-
-
-@api_view(["POST"])
-def grade(request: Request, problem_id: int, testcase_num: int):
-    code = request.data.get("code")
-    problem = Problem.objects.get(id=problem_id)  # TODO: Not Found Exception 처리
-    testcase: dict = problem.testcases[testcase_num - 1]  # TODO: IndexError 처리
-
-    tc_inp = testcase.get("input")
-    tc_outp = testcase.get("output")
-    timeout = testcase.get("timeout")
-    is_hidden = testcase.get("is_hidden")
-
-    file_path = BASE_DIR.joinpath("tmp", f"{uuid.uuid4()}.py")
-
-    # create directory if not exists
-    if not file_path.parent.exists():
-        file_path.parent.mkdir()
-
-    # write code to file
-    with open(file_path, "w") as f:
-        f.write(code)
-
-    outp, err = executor.run(file_path, code, tc_inp, timeout)
-
-    is_passed = str(outp).strip() == str(tc_outp).strip()
-
-    if is_hidden:
-        tc_inp = None
-        tc_outp = None
-        outp = None
-        err = None
-
-    # remove file
-    remove(file_path)
-
-    return Response(
-        {
-            "is_passed": is_passed,
-            "input": tc_inp,
-            "output": tc_outp,
-            "result": outp,
-            "error": err,
-        }
-    )
